+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file. The format 
 - Fix missing `ArbitraryOrderLagrangeElement.points` attribute.
 - Fix ignored mask `only_surface=True` for `RegionBoundary().mesh.cells_faces`.
 - Set default pressure to zero in `SolidBodyPressure()`.
+- Take the mesh from the global `x0`-field if `x0` is passed to `job.evaluate(x0=...)`.
 
 ## [5.2.0] - 2022-10-08